+++ pkg/kine/drivers/sqlite/driver_test.go
